+++ backend/src/main/java/com/taim/backend/service/payment/IPaymentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IPaymentService {
     List<Payment> getAllPayments();
     Payment savePayment(Payment payment);
-    Payment getPaymentById(Integer id);
+    Payment getPaymentById(Long id);
     void deletePayment(Payment payment);
     Payment updatePayment(Payment payment);
     Payment saveOrUpdatePayment(Payment payment);
