+++ backend/src/main/java/com/taim/backend/service/staff/IStaffService.java
@@ -14,7 +14,7 @@ public interface IStaffService {
     Staff saveStaff(Staff staff);
     Staff updateStaff(Staff staff);
     Staff getStaffByName(String name);
-    Staff getStaffById(Integer id);
+    Staff getStaffById(Long id);
     Staff saveOrUpdateStaff(Staff staff);
     void deleteStaff(Staff staff);
 }
