+++ model/src/main/java/com/taim/model/basemodels/BaseModel.java
@@ -16,7 +16,7 @@ public class BaseModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
     @Type(type = "org.jadira.usertype.dateandtime.joda.PersistentDateTime")
     @Column(name = "date_created" ,nullable = false)
     private DateTime dateCreated;
@@ -29,11 +29,11 @@ public class BaseModel {
 
     public BaseModel(){}
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
